// src/payload-types.ts
export interface User {
  id: number;
  role: 'admin' | 'client';
  name?: string | null;
  company?: string | null;
  inn?: string | null;
  phone?: string | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
// ...
export interface Application {
  id: number;
  title: string;
  serviceType: 'guarantee' | 'credit' | 'legal';
  status: 'review' | 'approved' | 'issued' | 'rejected';
  amount?: number | null;
  comment?: string | null;
  client: number | User;
  updatedAt: string;
  createdAt: string;
}